{{cookiecutter.repo_name}}/requirements.txt
modelyst-dbgen